+++ gusto_embedded/src/funcs/companyBenefitsDelete.ts
@@ -42,6 +42,8 @@ import { Result } from "../types/fp.js";
  *   - There are no payroll items associated with the company benefit
  *   - The benefit is not managed by a Partner or by Gusto (type must be 'External')
  *
+ * When the application has the `company_benefits:write:benefit_type_limited` data scope, the application can only delete company benefits for benefit types that are permitted for the application.
+ *
  * scope: `company_benefits:write`
  */
 export function companyBenefitsDelete(